Add simg check

+++ blueprint/feelpp/toolboxes_gen/upload/scripts/revert.sh
@@ -4,4 +4,6 @@ WORKDIR=$1
 URI=$2
 FILENAME=$3
 
-rm -f $WORKDIR/$FILENAME &> revert.log
+if [ -f $WORKDIR/$FILE ]; then
+    rm -f $WORKDIR/$FILENAME &> revert.log
+fi
